Conflicts:
	src/assets/i18n/en.json
	src/assets/i18n/es.json

+++ src/app/products/modals/confirmation-buy/confirmation-euclid/confirmation-euclid.component.ts
@@ -43,8 +43,11 @@ export class ConfirmationEuclidComponent implements OnInit {
   // list for File
   listFileBasket: Array<FileProductRequested> = new Array;
   listUrlFiles: Array<String> = new Array;
+  listFileLeftEye: Array<FileProductRequested> = new Array;
+  listFileRightEye: Array<FileProductRequested> = new Array;
   // boolean for delete file
   save_success: Boolean = false;
+  balance_modal: Boolean = false;
   company: Company = new Company();
   available: any;
   additional: any;
@@ -71,7 +74,7 @@ export class ConfirmationEuclidComponent implements OnInit {
   }
 
   close() {
-    if (!this.save_success) {
+    if (!this.save_success && !this.balance_modal) {
       this.listUrlFiles = this.buildUrlFiles();
       this.deleteAllFile();
     }
@@ -114,11 +117,12 @@ export class ConfirmationEuclidComponent implements OnInit {
   }
 
   save(): void {
-    this.spinner.show();
     if (this.typeBuy === 1) {
+      this.spinner.show();
       this.basketRequest.idUser = this.datos.idUser;
       this.basketRequest.productRequestedList = this.lista;
-      this.basketRequest.fileProductRequestedList = this.listFileBasket;
+      this.basketRequest.listFileRightEye = this.listFileRightEye;
+      this.basketRequest.listFileLeftEye = this.listFileLeftEye;
       this.basketService.saveBasket$(this.basketRequest).subscribe(res => {
         if (res.code === CodeHttp.ok) {
             this.save_success = true;
@@ -143,30 +147,33 @@ export class ConfirmationEuclidComponent implements OnInit {
       this.buyNow.idUser = this.datos.idUser;
       this.buyNow.productRequestedList = this.lista;
       this.buyNow.idRole = this.role;
-      this.buyNow.fileProductRequestedList = this.listFileBasket;
+      this.buyNow.listFileRightEye = this.listFileRightEye;
+      this.buyNow.listFileLeftEye = this.listFileLeftEye;
       this.validateAvailableBalance();
       if (this.available) {
-        this.orderService.saveOrderDirect$(this.buyNow).subscribe(res => {
-        if (res.code === CodeHttp.ok) {
-          this.save_success = true;
-          this.spinner.hide();
-          this.close();
-          this.translate.get('Order generated successfully', {value: 'Order generated successfully'}).subscribe(( res: string) => {
-            this.notification.success('', res);
-          });
-          this.redirectListOrder();
-        } else {
-          console.log(res);
-          this.translate.get('Connection Failed', { value: 'Connection Failed' }).subscribe((res: string) => {
-            this.notification.error('', res);
-          });
-          this.spinner.hide();
-          this.close();
-        }
-      }, error => {
-        console.log('error', error);
-      });
+          this.spinner.show();
+          this.orderService.saveOrderDirect$(this.buyNow).subscribe(res => {
+          if (res.code === CodeHttp.ok) {
+            this.save_success = true;
+            this.spinner.hide();
+            this.close();
+            this.translate.get('Order generated successfully', {value: 'Order generated successfully'}).subscribe(( res: string) => {
+              this.notification.success('', res);
+            });
+            this.redirectListOrder();
+          } else {
+            console.log(res);
+            this.translate.get('Connection Failed', { value: 'Connection Failed' }).subscribe((res: string) => {
+              this.notification.error('', res);
+            });
+            this.spinner.hide();
+            this.close();
+          }
+        }, error => {
+          console.log('error', error);
+        });
       } else {
+        this.balance_modal = true;
         this.openModal(); // No tiene disponible el balance de credito
         this.close();
       }
@@ -175,7 +182,11 @@ export class ConfirmationEuclidComponent implements OnInit {
 
   buildUrlFiles() {
     const listUrlFiles: Array<String> = new Array;
-    _.each(this.listFileBasket, function (file) {
+
+    _.each(this.listFileLeftEye, function (file) {
+      listUrlFiles.push(file.url);
+    });
+    _.each(this.listFileRightEye, function (file) {
       listUrlFiles.push(file.url);
     });
     return listUrlFiles;
@@ -240,8 +251,11 @@ export class ConfirmationEuclidComponent implements OnInit {
     const modalRef = this.modalService.open( NotificationBalanceComponent, { size: 'lg', windowClass: 'modal-content-border' });
     modalRef.componentInstance.buyNowModal = this.buyNow;
     modalRef.result.then((result) => {
+      this.save_success = true;
       this.ngOnInit();
     } , (reason) => {
+      this.save_success = true;
+      this.balance_modal = false;
       this.close();
     });
   }
